Added the nuke script.
The nuke script removes content that has not been licensed to other
developers for re-use. The content remains in the repository because it
ensures the application is buildable, but if any other developer wants
to have their own distribution (e.g., via a binary file), the
copyrighted material MUST be removed.

So the nuke script facilitates this.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>