Optimized the flood-fill algorithm.
Instead of using NSMutableArrays, which can dynamically allocate memory
up to arbitrary size, the algorithm is now using boolean arrays that are
preallocated with the maximum width and height of the images dimensions.
The boolean arrays default to false values. When a pixel is to be filled
with a color, the corresponding boolean value is set to true.

This is faster than using NSMutableArrays and ultimately uses much less
heap memory.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>