Fixed blend-mode bug that occurs when switching from fill to erase mode.
If you were to switch to erase mode during the middle of a flood-fill
op, the fill would occur in erase-mode. In other words, the fill would
actually erase the area it was meant to fill! This commit changes the
blend mode to normal when a flood-fill is actuall being added to the
canvas. Then it switches back to the last blend mode after the fill
operation is complete.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>